Separate journals and conferences in bibliography:
a) Conferences care about 'publisher'
b) Journals don't care about 'publisher'
c) Journals care about volume and number
d) Some journals are called "Transactions"
Example: IEEE Transactions on Circuits and Systems for Video Technology
e) Transactions and proceedings are usually about conferences and
symposiums, but not journals.
f) Some journals or conferences do not have continuous page numbering
but number per article instead.

IEEE and ACM related conferences usually have no "IEEE" and "ACM" in
title. However, the journals usually do. Since conferences always
have a publisher field, put "IEEE" or "ACM" (or other publisher) there
and remove from conference title. However, for the journal preserve
the entire name. Also, some journal names begin or end with the word
"journal", preserve it as well. Do not insert it if it is absent.

Also, when referring to conferences, do not write "The proceedings of
the second..." or "The transactions of the 4th...", it is very
inconsistent. Use the bare title of the conference.